Added encTitleKeys.bin parsing - nearly...!
The file is parsed, each block of titleid+key is looped through, but the
id and key are not extracted yet.
We need to extract them and then parse to the CreateTicket function
:hammer:
My python code from FunKeyCIA is shown for reference on what needs to be
done. Skipping system+update title ids... etc.